# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CarCraftr API",
    "description": "A comprehensive car information and community platform API",
    "version": "1.0.0",
    "contact": {
      "name": "CarCraftr Team",
      "email": "support@carcraftr.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.carcraftr.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/cars": {
      "get": {
        "summary": "Get cars with search and filters",
        "description": "Retrieve cars based on search query or individual filters",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search term for make, model, or year",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "make",
            "in": "query",
            "description": "Filter by car make",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "description": "Filter by car model",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "in": "query",
            "description": "Filter by car year",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of cars matching the criteria",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Car"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No cars found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "No cars found matching your search."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new car",
        "description": "Add a new car to the database",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["make", "model", "year"],
                "properties": {
                  "make": {
                    "type": "string",
                    "example": "Toyota"
                  },
                  "model": {
                    "type": "string",
                    "example": "Camry"
                  },
                  "year": {
                    "type": "string",
                    "example": "2023"
                  },
                  "specifications": {
                    "$ref": "#/components/schemas/CarSpecifications"
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "popularity": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Car created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
...
    "/api/news": {
      "get": {
        "summary": "Get all news",
        "description": "Retrieve all news articles sorted by publication date",
        "responses": {
          "200": {
            "description": "List of news articles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/News"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create news article",
        "description": "Add a new news article",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "description", "body"],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "body": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "publishedAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "News article created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/News"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/blogs": {
      "get": {
        "summary": "Get all blogs",
        "description": "Retrieve all blog posts sorted by creation date",
        "responses": {
          "200": {
            "description": "List of blog posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Blog"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create blog post",
        "description": "Add a new blog post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content", "author"],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "author": {
                    "type": "string"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blog post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Blog"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reviews": {
      "post": {
        "summary": "Create review",
        "description": "Add a new review for a car",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["carId", "rating", "title", "content"],
                "properties": {
                  "carId": {
                    "type": "string"
                  },
                  "rating": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "pros": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "cons": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Car": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Unique identifier for the car"
          },
          "make": {
            "type": "string",
            "description": "Car manufacturer",
            "example": "Toyota"
          },
          "model": {
            "type": "string",
            "description": "Car model",
            "example": "Camry"
          },
          "year": {
            "type": "string",
            "description": "Manufacturing year",
            "example": "2023"
          },
          "specifications": {
            "$ref": "#/components/schemas/CarSpecifications"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL to car image"
          },
          "popularity": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "description": "Popularity score"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CarSpecifications": {
        "type": "object",
        "properties": {
          "engine": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "V6"
              },
              "displacement": {
                "type": "string",
                "example": "3.5L"
              },
              "horsepower": {
                "type": "number",
                "example": 301
              },
              "torque": {
                "type": "number",
                "example": 267
              }
            }
          },
          "transmission": {
            "type": "string",
            "example": "8-speed automatic"
          },
          "drivetrain": {
            "type": "string",
            "example": "Front-wheel drive"
          },
          "fuelType": {
            "type": "string",
            "example": "Gasoline"
          },
          "seatingCapacity": {
            "type": "number",
            "example": 5
          },
          "cargoSpace": {
            "type": "string",
            "example": "15.1 cu ft"
          }
        }
      },
      "News": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "example": "New Electric Vehicle Launches"
          },
          "description": {
            "type": "string",
            "example": "Latest electric vehicle models hit the market"
          },
          "body": {
            "type": "string",
            "example": "Full article content..."
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Blog": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "example": "Car Maintenance Tips"
          },
          "content": {
            "type": "string",
            "example": "Blog post content..."
          },
          "author": {
            "type": "string",
            "example": "John Doe"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["maintenance", "tips", "cars"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Review": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "carId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "pros": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "helpfulCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 